# platform/docs/swagger.yaml
components:
  schemas:
...
    httpx.ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    httpx.Page-order_OrderHistoryView:
      properties:
        data:
          items:
            $ref: '#/components/schemas/order.OrderHistoryView'
          type: array
          uniqueItems: false
        limit:
          type: integer
        page:
          type: integer
        total:
          type: integer
      type: object
...
    order.CancelOrderRequest:
      properties:
        emiten:
          description: Emiten (listed instrument) code the order was placed on.
          example: BBCA
          type: string
        order_id:
          description: Id of the resting order to withdraw.
          example: 3
          format: int64
          minimum: 1
          type: integer
        participant:
          description: Participant (broker) code. Must be the broker that placed the
            order.
          example: YP
          type: string
      required:
      - emiten
      - order_id
      - participant
      type: object
    order.CancelOrderResponse:
      properties:
        order:
          $ref: '#/components/schemas/order.OrderView'
      type: object
    order.OrderHistoryView:
      properties:
        emiten:
          example: BBCA
          type: string
        id:
          example: 31
          format: int64
          type: integer
        participant:
          example: YP
          type: string
        price:
          description: 0 for a market order, which carries no price.
          example: 8000
          format: int64
          type: integer
        qty:
          example: 100
          format: int64
          type: integer
        remaining:
          description: Quantity not yet filled.
          example: 0
          format: int64
          type: integer
        seq:
          example: 31
          format: int64
          type: integer
        side:
          enum:
          - buy
          - sell
          example: buy
          type: string
        status:
          enum:
          - open
          - filled
          - cancelled
          example: filled
          type: string
        type:
          enum:
          - limit
          - market
          example: limit
          type: string
      type: object
    order.OrderView:
      properties:
        id:
          description: Database id of the order.
          example: 3
          format: int64
          type: integer
        remaining:
          description: Quantity not yet filled.
          example: 0
          format: int64
          type: integer
        status:
          description: Lifecycle state after matching.
          enum:
          - open
          - filled
          - cancelled
          example: filled
          type: string
      type: object
    order.SubmitOrderRequest:
      properties:
        emiten:
          description: Emiten (listed instrument) code.
          example: BBCA
          type: string
        participant:
          description: Participant (broker) code.
          example: YP
          type: string
        price:
          description: |-
            Whole rupiah, never a float. Required and > 0 for a limit order; ignored
            and stored as 0 for a market order.
          example: 8000
          format: int64
          type: integer
        qty:
          description: Quantity to trade. Must be > 0.
          example: 100
          format: int64
          minimum: 1
          type: integer
        side:
          description: Order direction.
          enum:
          - buy
          - sell
          example: buy
          type: string
        type:
          description: Order type. A market order never rests in the book.
          enum:
          - limit
          - market
          example: limit
          type: string
      required:
      - emiten
      - participant
      - qty
      - side
      - type
      type: object
    order.SubmitOrderResponse:
      properties:
        order:
          $ref: '#/components/schemas/order.OrderView'
        trades:
          items:
            $ref: '#/components/schemas/order.TradeView'
          type: array
          uniqueItems: false
      type: object
    order.TradeView:
      properties:
        buy_order_id:
          description: Id of the buy side of this execution.
          example: 3
          format: int64
          type: integer
        price:
          description: Execution price, which is always the passive (resting) order's
            price.
          example: 8000
          format: int64
          type: integer
        qty:
          description: Quantity executed.
          example: 100
          format: int64
          type: integer
        sell_order_id:
          description: Id of the sell side of this execution.
          example: 1
          format: int64
          type: integer
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: Static administrative key, from configuration. Guards every /api/admin route.
      in: header
      name: X-API-Key
      type: apiKey
    ParticipantKeyAuth:
      description: Per-broker key, issued by an admin and stored hashed. Guards every /api/participant route. Returned once at issue and never retrievable again.
      in: header
      name: X-Participant-Key
      type: apiKey
...
info:
  description: |-
    REST surface for the JAST core matching engine (JATS-style continuous
    matching for Bursa Efek Indonesia). The engine knows only brokers
    (participants), not individual investors, and stops at trade execution.


    There are two authentication tiers, and they never mix:

    - **Participant** (`/api/participant/*`, `/ws/participant/*`) — a broker
    authenticates with its own key, stored hashed in the database and sent as
    `X-Participant-Key`. Keys are issued and revoked per broker at runtime.
    - **Admin** (`/api/admin/*`, `/ws/admin/*`) — the single static key from
    configuration, sent as `X-API-Key`.

    Orders enter only through POST /api/participant/orders. The order book can
    be polled or streamed over WebSocket (documented, but not exercisable from
    Swagger UI, since browsers cannot set headers on a WS handshake).
  title: JAST Core API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /api/admin/orders:
    get:
      description: |-
        Every order ever submitted, newest first, optionally filtered by emiten,
        participant, or status. This is the audit trail; the live book is served
        by the orderbook endpoints.
      operationId: listOrders
      parameters:
      - description: Filter by emiten code
        example: BBCA
        in: query
        name: emiten
        schema:
          type: string
      - description: Filter by broker code
        example: YP
        in: query
        name: participant
        schema:
          type: string
      - description: Filter by status
        in: query
        name: status
        schema:
          enum:
          - open
          - filled
          - cancelled
          type: string
      - description: Page number, from 1
        in: query
        name: page
        schema:
          default: 1
          minimum: 1
          type: integer
      - description: Items per page, max 100
        in: query
        name: limit
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Page-order_OrderHistoryView'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Unknown emiten/participant, or bad status
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
      security:
      - ApiKeyAuth: []
      summary: List order history
      tags:
      - admin
...
  /api/participant/orders:
    post:
      description: |-
        Runs the order through continuous matching and returns the trades it
        produced. A limit order with leftover quantity rests in the book (open);
        a market order with leftover quantity is cancelled and never booked.
        The order, its trades and the fills against the resting orders it
        consumed are persisted in a single transaction.

        A sell is rejected if the broker does not hold enough shares once its
        resting sell orders are counted, so a position can never go negative.

        **Auto-rejection.** A limit order priced outside the instrument's session
        band is rejected and never reaches the book, so the price it asked for
        never becomes a level anything else can match against. The band is
        `reference_price ± emiten_halt_bps` (30% by default) and is anchored to the
        session reference, not to the last trade — an anchor that moved with the
        market would let the price walk out of the band one legal step at a time.
        The rejection message quotes the permitted range.

        **Circuit breaker.** A trade that prints at either edge of the band halts
        the instrument for halt_duration_seconds (2 minutes by default). Orders
        submitted during a halt are rejected with the resume time; the book stays
        readable and resting orders keep their time priority. Trading reopens by
        itself — no request is needed.
      operationId: submitOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/order.SubmitOrderRequest'
              description: Order to submit
              summary: body
        description: Order to submit
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order.SubmitOrderResponse'
          description: 'Order processed: the resulting order state and any trades'
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Unknown emiten/participant, bad side/type, qty <= 0, price
            <= 0 for a limit order, insufficient shares, an inactive emiten, a price
            outside the permitted band, or a halted instrument
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-Participant-Key
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Persistence failure
      security:
      - ParticipantKeyAuth: []
      summary: Submit a new order
      tags:
      - participant
  /api/participant/orders/cancel:
    post:
      description: |-
        Withdraws an order that is still resting in the book and releases
        what it had reserved — shares for a sell, cash for a buy — so the
        broker can use them again immediately. The quantity that already
        filled stays filled; only the remainder is withdrawn.

        Only the broker that placed the order may cancel it. An order that
        has already filled or been cancelled is not resting, so it cannot be
        withdrawn and the request is rejected.

        The book row and the stored order are cancelled in one step, and the
        updated book is published to subscribers. A cancel produces no trade
        and therefore moves no price.

        Cancelling is allowed while the instrument is halted: a halt stops new
        orders from arriving, it does not trap the ones already resting.
      operationId: cancelOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/order.CancelOrderRequest'
              description: Order to cancel
              summary: body
        description: Order to cancel
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order.CancelOrderResponse'
          description: 'Order cancelled: its final state and the quantity released'
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Unknown emiten/participant, order_id <= 0, an order that is
            not resting, or an order belonging to another participant
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-Participant-Key
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Persistence failure
      security:
      - ParticipantKeyAuth: []
      summary: Cancel a resting order
      tags:
      - participant
...
servers:
- description: Local development
  url: http://localhost:8080